working but not working

+++ script.js
@@ -1,16 +1,32 @@
+// Разобраться с last_id когда теряем фокус элемента
+// При переходе к заметке непосредственно через ссылку, нужно считать ее, проверить есть ли такой id в localStorage
+// Если такого значения в хранилище нету, pushToURL('undefined') unselect(), 
 
-// TODO: сделать рутинг, в переменной last_id хранится последний id элемента на который клинкули перед тем как рефрешнуть страницу. 
-// Идея в том, что бы использовать last_id как часть ссылки на заметку.
-// Когда пользователь переключает фокус на другую заметку, ее id записывается в переменную last_id, если last_id != null,
-// то при загрузке страницы мы переходим по ссылке содержащей last_id (но при загрузке last_id никогда не будет пустым oh shiiiieeeeeeet..)
+const pushToURL = (name, key, value) => window.history.replaceState(null, null, `${name}?${key}=${value}`)
 
 var cur_id;
 var notes = [];
-if(localStorage.length != 0 ) {
+if(localStorage.length > 1 ) {
     notes = JSON.parse(localStorage.getItem('notes'));
 }
-console.log(notes);
 
+console.log(notes); 
+
+//cur = Number(((window.location.href).split('='))[1]);
+
+var isInStorage = false;
+
+for (let i = 0; i < notes.length; i++) {
+    if (localStorage.getItem('last-id') == notes[i]) {
+        pushToURL('', 'id', localStorage.getItem('last-id'));
+        isInStorage = true;
+    }
+}
+
+if (!isInStorage) {
+    localStorage.setItem('last-id', 'undefined');
+    pushToURL('', 'id', 'undefined');
+}
 
 
 function parseFunction() {
@@ -21,7 +37,6 @@ function parseFunction() {
         temp_arr = JSON.parse(localStorage.getItem(notes[i]));
         console.log(temp_arr);
         
-
         let note = document.createElement('div');
         note.className =  "note";
         note.id = notes[i];
@@ -40,17 +55,22 @@ function parseFunction() {
         var parentNode = note.parentNode;
         parent.insertBefore(note, parentNode);
     }
+    
+
+    if (localStorage.getItem('last-id') != 'undefined') {
+        focusNote(localStorage.getItem('last-id'));
+    }
 }
 
 function addNote(name = "Blank note", id = String(Math.random()).slice(2,8), text = '') {
     var date = new Date();
     var time = 'Last change: Hours: ' +date.getHours() + ', Minutes: ' + date.getMinutes() + ', Seconds: ' + date.getSeconds();
     unselect();
-
-
+    localStorage.setItem('last-id','undefined');
+    
     document.getElementById("text-area-input").setAttribute('readonly','true');
     notes.splice(0,0,id);
-    console.log(notes);
+    //console.log(notes);
     localStorage.setItem('notes', JSON.stringify(notes));
 
 
@@ -81,10 +101,19 @@ function addNote(name = "Blank note", id = String(Math.random()).slice(2,8), tex
     document.location.reload();
 }
 
-function focusNote() {
+function focusNote(cur) {
+    if (cur == null) {
+        unselect();
+        cur_id = document.activeElement.id;
     
-    unselect();
-    cur_id = document.activeElement.id;
+    } else {
+        unselect();
+        cur_id = cur;
+    }
+   
+    console.log(cur_id);
+
+    pushToURL('', 'id', cur_id);
 
     localStorage.setItem('last-id', cur_id);
 
@@ -106,6 +135,8 @@ function focusNote() {
     document.getElementById(cur_id).querySelector('div.note-header input[id=note-header-text]').value = name;
     document.getElementById("text-area-input").value = text;
     document.getElementById("time-section").value = time;
+
+    document.getElementById("text-area-input").focus();
 }
 
 function update() {
@@ -121,29 +152,37 @@ function update() {
     localStorage.setItem(cur_id,JSON.stringify([name,content,time] ));
 
     if (cur_id == Number(notes[0])) {
-
+        
     } else {
         var temp_indx = notes.indexOf(cur_id);
         notes.splice(temp_indx, 1)
         notes.splice(0, 0, cur_id)
         console.log(notes);
         localStorage.setItem('notes', JSON.stringify(notes));
+        reloadPage();
     }   
 }
 
 function unselect() {
-    if(cur_id == null) {
+    if(cur_id == undefined) {
         return;
     }
     document.getElementById(cur_id).classList.remove('selected-note');
 }
 
 function deleteNote() {
-    var temp_indx = notes.indexOf(cur_id);
-    notes.splice(temp_indx, 1);
-    localStorage.setItem('notes',  JSON.stringify(notes));
-    localStorage.removeItem(cur_id);
-    document.location.reload();
+    if (confirm('Are you sure?')) {
+        var temp_indx = notes.indexOf(cur_id);
+        notes.splice(temp_indx, 1);
+        localStorage.setItem('notes',  JSON.stringify(notes));
+        localStorage.removeItem(cur_id);
+        localStorage.setItem('last-id', 'undefined');
+        console.log(localStorage.getItem('last-id'));
+        window.location.href = window.location.origin + window.location.pathname +'?id=undefined';
+        document.location.reload();
+    }
+   
+   
 }
 
 function reloadPage() {
